Move dompurify to browser layer
