+++ src/vs/base/browser/dompurify/cgmanifest.json
