+++ src/vs/base/browser/dompurify/dompurify.d.ts
