+++ src/vs/base/browser/dompurify/dompurify.js
