+++ src/vs/base/browser/dompurify/dompurify.license.txt
